Chunk No needs to be long.

+++ Messages.cs
@@ -3,7 +3,7 @@ using Akka.Actor;
 namespace reS3m.Messages
 {
     public struct Chunk {
-        public int No;
+        public long No;
         public long Start;
         public long End;
         public long Size => End - Start + 1;
@@ -29,9 +29,9 @@ namespace reS3m.Messages
     public class ChunkFlushed {
         public IActorRef Actor {get; }
         public string S3ObjectName {get;}
-        public int ChunkNo {get;}
+        public long ChunkNo {get;}
 
-        public ChunkFlushed (IActorRef actor, string s3obj, int chunkNo) {
+        public ChunkFlushed (IActorRef actor, string s3obj, long chunkNo) {
             Actor = actor;
             S3ObjectName = s3obj;
             ChunkNo = chunkNo;
@@ -63,9 +63,9 @@ namespace reS3m.Messages
 
     public class FlushChunk {
         public string S3ObjectName {get;}
-        public int ChunkNo {get;}
+        public long ChunkNo {get;}
 
-        public FlushChunk(string s3ObjectName, int chunkNo)
+        public FlushChunk(string s3ObjectName, long chunkNo)
         {
             S3ObjectName = s3ObjectName;
             ChunkNo = chunkNo;
